GurobiOpts: An S4 class to represent Gurobi parameters

This class is used to store Gurobi input parameters.

Slots

Threads

integer number of cores to use for processing. Defaults to 1L.

MIPGap

numeric MIP gap specifying minimum solution quality. Defaults to 0.1.

Method

integer Algorithm to use for solving model. Defaults to 0L.

Presolve

integer code for level of computation in presolve. Defaults to 2.

TimeLimit

integer number of seconds to allow for solving. Defaults to NA_integer_, and so a time limit is not imposed.

NumberSolutions

integer number of solutions to generate. Defaults to 1L.

MultipleSolutionsMethod

integer name of method to obtain multiple solutions (used when NumberSolutions is greater than one). Available options are "benders.cuts", "solution.pool.0", "solution.pool.1", and "solution.pool.2". The "benders.cuts" method produces a set of distinct solutions that are all within the optimality gap. The "solution.pool.0" method returns all solutions identified whilst trying to find a solution that is within the specified optimality gap. The "solution.pool.1" method finds one solution within the optimality gap and a number of additional solutions that are of any level of quality (such that the total number of solutions is equal to number_solutions). The "solution.pool.2" finds a specified number of solutions that are nearest to optimality. The search pool methods correspond to the parameters used by the Gurobi software suite (see http://www.gurobi.com/documentation/8.0/refman/poolsearchmode.html#parameter:PoolSearchMode). Defaults to "benders.cuts".

See also