Create RapResults object

This function creates a new RapResults object.

RapResults(summary, selections, amount.held, space.held, logging.file,
  .cache = new.env())

Arguments

summary
data.frame with summary information on solutions. See details below for more information.

selections
matrix with binary selections. The cell \(x_{ij}\) denotes if planning unit \(j\) is selected in the \(i\)'th solution.

amount.held
matrix with the amount held for each species in each solution.

space.held
matrix with the proportion of attribute space sampled for each species in each solution.

logging.file
character Gurobi log files.

.cache
environment used to cache calculations.

Value

RapResults object

Details

The summary table follows Marxan conventions (summary.dat in http://marxan.net/downloads/uq_marxan_web_2/module5.html). The columns are:

Run_Number

The index of each solution in the object.

Status

The status of the solution. The values in this column correspond to outputs from the Gurobi software package (http://www.gurobi.com/documentation/6.5/refman/optimization_status_codes.html).

Score

The objective function for the solution.

Cost

Total cost associated with a solution

Planning_Units

Number of planning units selected in a solution

Connectivity_Total

The total amount of shared boundary length between all planning units. All solutions in the same object should have equal values for this column.

Connectivity_In

The amount of shared boundary length among planning units selected in the solution

Connectivity_Edge

The amount of exposed boundary length in the solution

Connectivity_Out

The number of shared boundary length among planning units not selected in the solution

Connectivity_Fraction

The ratio of shared boundary length in the solution (Connectivity_In) to the total amount of boundary length (Connectivity_Edge). This ratio is an indicator of solution quality. Solutions with a lower ratio will have less planning units and will be more efficient.

Note

slot best is automatically determined based on data in summary.

See also

RapResults-class read.RapResults.