RapResults: An S4 class to represent RAP results

This class is used to store RAP results.

Details

The summary table follows Marxan conventions (https://marxansolutions.org/). The columns are:

Run_Number

The index of each solution in the object.

Status

The status of the solution. The values in this column correspond to outputs from the Gurobi software package (http://www.gurobi.com/documentation/6.5/refman/optimization_status_codes.html).

Score

The objective function for the solution.

Cost

Total cost associated with a solution.

Planning_Units

Number of planning units selected in a solution.

Connectivity_Total

The total amount of shared boundary length between all planning units. All solutions in the same object should have equal values for this column.

Connectivity_In

The amount of shared boundary length among planning units selected in the solution.

Connectivity_Edge

The amount of exposed boundary length in the solution.

Connectivity_Out

The number of shared boundary length among planning units not selected in the solution.

Connectivity_Fraction

The ratio of shared boundary length in the solution (Connectivity_In) to the total amount of boundary length (Connectivity_Edge). This ratio is an indicator of solution quality. Solutions with a lower ratio will have less planning units and will be more efficient.

Slots

summary

base::data.frame() with summary information on solutions.

selections

base::matrix() with binary selections. The cell \(x_{ij}\) denotes if planning unit \(j\) is selected in the \(i\)'th solution.

amount.held

base::matrix() with the amount held for each species in each solution.

space.held

base::matrix() with the proportion of attribute space sampled for each species in each solution.

best

integer with index of best solution.

logging.file

character Gurobi log files.

.cache

base::environment() used to store extra data.

See also