Case-study dataset for a conservation planning exercise

This dataset contains data to generate example prioritisations for the pale-headed Rosella (Platycercus adscitus) in Queensland, Australia. Specific objects in the dataset include:

cs_pus: SpatialPolygonsDataFrame planning units. The units were generated as \(30km^2\) squares across the species' range, and then clipped to the Queensland, Australia (using data obtained from the http://www.abs.gov.au/ausstats/abs@.nsf/mf/1259.0.30.001?OpenDocument). They were then overliad with Australia's protected area network (obtained from the World Database on Protected Areas (WDPA) at http://www.protectedplanet.net/). This attribute table has 3 fields. The 'area' field contains the units' area, the 'cost' field is set to 1 for all units, and the 'status' filed indicates if 50% or more of the units' extent is covered by protected areas.

cs_spp: RasterLayer-class probability distribution map for the P. adscitus clipped to Queensland, Australia. This map was derived from records obtained from The Atlas of Living Australia (http://ala.org.au/).

cs_space: RasterStack describing broad-scale variation across Queensland. containing the projections of the first two axes from a annual precipitation data across the species range (obtained from http://www.worldclim.org/ as BIO12, and resampled to \(10km^2\) resolution).

cs_pus

cs_space

cs_spp

Format

cs_pus: SpatialPolygonsDataFrame, cs_spp: RasterLayer-class, cs_space: RasterStack-class.

Examples

# load data
data(cs_pus, cs_spp, cs_space)
# plot data
not_run({
  plot(cs_pus)
  plot(cs_spp)
  plot(cs_space)
})