Log file

This function returns the Gurobi log file (*.log) associated with solving an optimization problem.

logging.file(x, y)

# S3 method for RapResults
logging.file(x, y = 0)

# S3 method for RapSolved
logging.file(x, y = 0)

Arguments

x

RapResults() or RapSolved() object.

y

Available inputs include: NULL to return all values, integer number specifying the solution for which the log file should be returned, and 0 to return log file for the best solution.

Note

The term logging file was used due to collisions with the log function.

See also

Examples

# \dontrun{
# load data
data(sim_rs)

# log file for the best solution
cat(logging.file(sim_rs, 0))
#> 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:41:38 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30306 rows, 30100 columns and 120000 nonzeros
#> Model fingerprint: 0xfe534681
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 827 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 13.37s
#> Presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Use crossover to convert LP symmetric solution to basic solution...
#> 
#> Root crossover log...
#> 
#>      511 PPushes remaining with PInf 0.0000000e+00                14s
#>        0 PPushes remaining with PInf 0.0000000e+00                14s
#> 
#>   Push phase complete: Pinf 0.0000000e+00, Dinf 0.0000000e+00     14s
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 2464 iterations, 0.98 seconds (0.64 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   14s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   15s
#> 
#> Explored 1 nodes (5366 simplex iterations) in 15.01 seconds (9.00 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000%

# log file for the second solution
cat(logging.file(sim_rs, 2))
#> 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:41:38 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30306 rows, 30100 columns and 120000 nonzeros
#> Model fingerprint: 0xfe534681
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 827 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 13.37s
#> Presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Use crossover to convert LP symmetric solution to basic solution...
#> 
#> Root crossover log...
#> 
#>      511 PPushes remaining with PInf 0.0000000e+00                14s
#>        0 PPushes remaining with PInf 0.0000000e+00                14s
#> 
#>   Push phase complete: Pinf 0.0000000e+00, Dinf 0.0000000e+00     14s
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 2464 iterations, 0.98 seconds (0.64 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   14s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   15s
#> 
#> Explored 1 nodes (5366 simplex iterations) in 15.01 seconds (9.00 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000%
#> 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:41:54 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30307 rows, 30100 columns and 120100 nonzeros
#> Model fingerprint: 0x7b9a1818
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 829 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 12.13s
#> Presolved: 29462 rows, 30083 columns, 125689 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29462 rows, 30083 columns, 125689 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     13s
#>     4020    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     4020    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 4020 iterations, 1.60 seconds (1.94 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   14s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   14s
#> 
#> Explored 1 nodes (7363 simplex iterations) in 14.62 seconds (10.43 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000%

# log files for all solutions
cat(logging.file(sim_rs, NULL))
#> 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:41:38 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30306 rows, 30100 columns and 120000 nonzeros
#> Model fingerprint: 0xfe534681
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 827 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 13.37s
#> Presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Use crossover to convert LP symmetric solution to basic solution...
#> 
#> Root crossover log...
#> 
#>      511 PPushes remaining with PInf 0.0000000e+00                14s
#>        0 PPushes remaining with PInf 0.0000000e+00                14s
#> 
#>   Push phase complete: Pinf 0.0000000e+00, Dinf 0.0000000e+00     14s
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 2464 iterations, 0.98 seconds (0.64 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   14s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   15s
#> 
#> Explored 1 nodes (5366 simplex iterations) in 15.01 seconds (9.00 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000% 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:41:38 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30306 rows, 30100 columns and 120000 nonzeros
#> Model fingerprint: 0xfe534681
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 827 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 13.37s
#> Presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Use crossover to convert LP symmetric solution to basic solution...
#> 
#> Root crossover log...
#> 
#>      511 PPushes remaining with PInf 0.0000000e+00                14s
#>        0 PPushes remaining with PInf 0.0000000e+00                14s
#> 
#>   Push phase complete: Pinf 0.0000000e+00, Dinf 0.0000000e+00     14s
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 2464 iterations, 0.98 seconds (0.64 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   14s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   15s
#> 
#> Explored 1 nodes (5366 simplex iterations) in 15.01 seconds (9.00 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000%
#> 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:41:54 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30307 rows, 30100 columns and 120100 nonzeros
#> Model fingerprint: 0x7b9a1818
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 829 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 12.13s
#> Presolved: 29462 rows, 30083 columns, 125689 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29462 rows, 30083 columns, 125689 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     13s
#>     4020    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     4020    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 4020 iterations, 1.60 seconds (1.94 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   14s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   14s
#> 
#> Explored 1 nodes (7363 simplex iterations) in 14.62 seconds (10.43 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000% 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:41:38 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30306 rows, 30100 columns and 120000 nonzeros
#> Model fingerprint: 0xfe534681
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 827 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 13.37s
#> Presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29461 rows, 30083 columns, 125589 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     1950    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Use crossover to convert LP symmetric solution to basic solution...
#> 
#> Root crossover log...
#> 
#>      511 PPushes remaining with PInf 0.0000000e+00                14s
#>        0 PPushes remaining with PInf 0.0000000e+00                14s
#> 
#>   Push phase complete: Pinf 0.0000000e+00, Dinf 0.0000000e+00     14s
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     2464    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 2464 iterations, 0.98 seconds (0.64 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   14s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   15s
#> 
#> Explored 1 nodes (5366 simplex iterations) in 15.01 seconds (9.00 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000%
#> 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:41:54 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30307 rows, 30100 columns and 120100 nonzeros
#> Model fingerprint: 0x7b9a1818
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 829 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 12.13s
#> Presolved: 29462 rows, 30083 columns, 125689 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29462 rows, 30083 columns, 125689 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     13s
#>     4020    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     4020    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 4020 iterations, 1.60 seconds (1.94 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   14s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   14s
#> 
#> Explored 1 nodes (7363 simplex iterations) in 14.62 seconds (10.43 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000%
#> 
#> Gurobi 10.0.0 (linux64, R) logging started Fri 23 Dec 2022 02:42:08 PM NZDT
#> 
#> Set parameter Username
#> Set parameter MIPGap to value 0.1
#> Set parameter Method to value 0
#> Set parameter LogFile to value "/tmp/RtmpJ30SOp/file57617b549888.log"
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Academic license - for non-commercial use only - expires 2023-12-18
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (linux64)
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Optimize a model with 30308 rows, 30100 columns and 120200 nonzeros
#> Model fingerprint: 0x8e215663
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99.0000000
#> Presolve removed 829 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 845 rows and 15 columns (presolve time = 10s) ...
#> Presolve removed 845 rows and 17 columns
#> Presolve time: 12.02s
#> Presolved: 29463 rows, 30083 columns, 125789 nonzeros
#> Variable types: 0 continuous, 30083 integer (30083 binary)
#> Root relaxation presolved: 29463 rows, 30083 columns, 125789 nonzeros
#> 
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02     13s
#>     5394    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#>     5394    2.0000000e+01   0.000000e+00   0.000000e+00     14s
#> 
#> Root relaxation: objective 2.000000e+01, 5394 iterations, 2.14 seconds (2.93 work units)
#> Total elapsed time = 16.39s
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> H    0     0                      20.0000000    0.00000   100%     -   16s
#>      0     0          -    0        20.00000   20.00000  0.00%     -   16s
#> 
#> Explored 1 nodes (12421 simplex iterations) in 16.51 seconds (13.25 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000%
# }