raptr: Representative and Adequate Prioritisation Toolkit in R

Biodivesity is in crisis. The overarching aim of conservation is to preserve biodiversity patterns and processes. To this end, protected areas are established to buffer species and preserve biodiversity processes. But resources are limited and so protected areas must be cost-effective. This package contains tools to generate plans for protected areas (prioritizations). Conservation planning data are used to construct an optimisation problem, which is then solved to yield prioritizations. To solve the optimisation problems in a feasible amount of time, this package uses the commercial 'Gurobi' software package (obtained from <http://www.gurobi.com/>). Additionally, the 'rgurobi' package can also be installed to provide extra functionality (obtained from <http://github.com/jeffreyhanson/rgurobi>). For more information on using this package, see Hanson et al. (2017) <doi:10.1111/2041-210X.12862>.

Details

The main classes used in this package are used to store input data and prioritisations:

GurobiOpts-class

parameters for solving optimisation problems using Gurobi.

RapReliableOpts-class

parameters for the reliable formulation of RAP.

RapUnreliableOpts-class

parameters for the unreliable formulation of RAP.

RapData-class

planning unit, species data, and demand points for RAP.

RapUnsolved-class

contains all the data and input parameters required to generate prioritisations using RAP. This class contains a GurobiOpts-class object, a RapReliableOpts-class or RapUnreliableOpts-class object, and a RapData-class object.

RapResults-class

prioritisations and summary statistics on their performance.

RapSolved-class

contains all the input data, parameters and output data. This class contains all the objects in a RapUnsolved object and also a RapResults-class object.

Type vignette("raptr") for a tutorial on how to use this package.