Summary of solutions

Extracts summary of solutions in a RapResults or RapSolved object.

Arguments

object
RapResults, or RapSolved object.

...
not used.

Value

data.frame

Details

This table follows Marxan conventions (summary.dat in http://marxan.net/downloads/uq_marxan_web_2/module5.html). The columns are:

Run_Number

The index of each solution in the object.

Status

The status of the solution. The values in this column correspond to outputs from the Gurobi software package (http://www.gurobi.com/documentation/6.5/refman/optimization_status_codes.html).

Score

The objective function for the solution.

Cost

Total cost associated with a solution

Planning_Units

Number of planning units selected in a solution

Connectivity_Total

The total amount of shared boundary length between all planning units. All solutions in the same object should have equal values for this column.

Connectivity_In

The amount of shared boundary length among planning units selected in the solution

Connectivity_Edge

The amount of exposed boundary length in the solution

Connectivity_Out

The number of shared boundary length among planning units not selected in the solution

Connectivity_Fraction

The ratio of shared boundary length in the solution (Connectivity_In) to the total amount of boundary length (Connectivity_Edge). This ratio is an indicator of solution quality. Solutions with a lower ratio will have less planning units and will be more efficient.

See also

RapResults, RapSolved.

Examples

# load data
data(sim_rs)
# show summary
summary(sim_rs)

#>   Run_Number  Status Score Cost Planning_Units Connectivity_Total
#> 1          1 OPTIMAL    20   20             20                220
#> 2          2 OPTIMAL    20   20             20                220
#> 3          3 OPTIMAL    20   20             20                220
#>   Connectivity_In Connectivity_Edge Connectivity_Out Connectivity_In_Fraction
#> 1               8               142               70               0.03636364
#> 2               3               141               76               0.01363636
#> 3               7               141               72               0.03181818