Convert SpatialPolygons to PolySet data

This function converts spatial sp::SpatialPolygons() and sp::SpatialPolygonsDataFrame() objects to PBSmapping::PolySet() objects.

SpatialPolygons2PolySet(x, n_preallocate)

# S3 method for SpatialPolygonsDataFrame
SpatialPolygons2PolySet(x, n_preallocate = 10000L)

# S3 method for SpatialPolygons
SpatialPolygons2PolySet(x, n_preallocate = 10000L)

Arguments

x

sp::SpatialPolygons() or sp::SpatialPolygonsDataFrame() object.

n_preallocate

integer How much memory should be preallocated for processing? Ideally, this number should equal the number of vertices in the sp::SpatialPolygons() object. If data processing is taking too long consider increasing this value.

Value

PBSmapping::PolySet() object.

Note

Be aware that this function is designed to be as fast as possible, but as a result it depends on C++ code and if used inappropriately this function will crash R.

See also

Examples

# generate SpatialPolygons object
sim_pus <- sim.pus(225L)

# convert to PolySet
x <- SpatialPolygons2PolySet(sim_pus)