Generate demand points for RAP

This function generates demand points to characterize a distribution of points.

make.DemandPoints(points, n = 100L, quantile = 0.5,
  kernel.method = c("ks", "hypervolume")[1], ...)

Arguments

points
matrix object containing points.

n
integer number of demand points to use for each attribute space for each species. Defaults to 100L.

quantile
numeric quantile to generate demand points within. If 0 then demand points are generated across the full range of values the points intersect. Defaults to 0.5.

kernel.method
character name of kernel method to use to generate demand points. Defaults to 'ks'.

...
arguments passed to kernel density estimating functions

Value

DemandPoints object.

Details

Broadly speaking, demand points are generated by fitting a kernal to the input points. A shape is then fit to the extent of the kernal, and then points are randomly generated inside the shape. The demand points are generated as random points inside the shape. The weights for each demand point are calculated the estimated density of input points at the demand point. By supplying 'ks' as an argument to method in kernel.method, the shape is defined using a minimum convex polygon mcp and kde is used to fit the kernel. Note this can only be used when the data is low-dimensional (d < 3). By supplying "hypervolume" as an argument to method, the hypervolume function is used to create the demand points. This method can be used for hyper-dimensional data (\(d << 3\)).

See also

Examples

# set random number generator seed
set.seed(500)

# load data
data(cs_spp, cs_space)

# generate species points
species.points <- randomPoints(cs_spp[[1]], n = 100, prob = TRUE)
env.points <- raster::extract(cs_space, species.points)

# generate demand points for a 1d space using ks
dps1 <- make.DemandPoints(points = env.points[, 1], kernel.method = "ks")

# print object
print(dps1)

#> An object of class "DemandPoints"
#> Slot "coords":
#>                 [,1]
#>   [1,]  0.0808103965
#>   [2,] -0.0046110220
#>   [3,]  0.0668303872
#>   [4,] -0.0041704372
#>   [5,]  0.0967735148
#>   [6,]  0.0861980940
#>   [7,]  0.0132235371
#>   [8,]  0.0310942531
#>   [9,]  0.0825810501
#>  [10,] -0.0060209785
#>  [11,]  0.0564411165
#>  [12,]  0.0481650725
#>  [13,]  0.0344428694
#>  [14,]  0.0926038712
#>  [15,] -0.0147107140
#>  [16,]  0.0756575648
#>  [17,]  0.0369092063
#>  [18,]  0.0799855539
#>  [19,] -0.0117221542
#>  [20,]  0.0577459375
#>  [21,]  0.0920794479
#>  [22,]  0.0952088934
#>  [23,] -0.0037990803
#>  [24,]  0.0305098018
#>  [25,]  0.0693633653
#>  [26,]  0.0298491904
#>  [27,] -0.0151671063
#>  [28,]  0.0661814398
#>  [29,]  0.0964706920
#>  [30,]  0.0960502611
#>  [31,]  0.0230169779
#>  [32,] -0.0062315867
#>  [33,]  0.0466713173
#>  [34,]  0.0430020562
#>  [35,] -0.0034828316
#>  [36,]  0.0139532329
#>  [37,]  0.0610430386
#>  [38,]  0.0638028609
#>  [39,]  0.0141869346
#>  [40,]  0.0614487429
#>  [41,]  0.0174622277
#>  [42,]  0.0569979267
#>  [43,]  0.0758035877
#>  [44,]  0.0911338269
#>  [45,]  0.0613651952
#>  [46,]  0.0712111251
#>  [47,] -0.0153747193
#>  [48,]  0.0053567867
#>  [49,]  0.0001290641
#>  [50,]  0.0748831151
#>  [51,]  0.0913816999
#>  [52,]  0.0696987094
#>  [53,]  0.0081708489
#>  [54,] -0.0165420546
#>  [55,]  0.0685402103
#>  [56,]  0.0458259356
#>  [57,]  0.0473664530
#>  [58,]  0.0598048637
#>  [59,]  0.0029895715
#>  [60,]  0.0134756199
#>  [61,]  0.0898856664
#>  [62,]  0.0962462809
#>  [63,]  0.0379707443
#>  [64,]  0.0634802594
#>  [65,]  0.0808849064
#>  [66,]  0.0236644932
#>  [67,]  0.0044193706
#>  [68,]  0.0343277188
#>  [69,]  0.0193473450
#>  [70,]  0.0633850620
#>  [71,] -0.0002382388
#>  [72,]  0.0894711454
#>  [73,]  0.0148280822
#>  [74,]  0.0381031305
#>  [75,] -0.0069926641
#>  [76,]  0.0202786370
#>  [77,]  0.0664897503
#>  [78,] -0.0106829767
#>  [79,]  0.0283130895
#>  [80,]  0.0461905186
#>  [81,]  0.0069314659
#>  [82,]  0.0376396247
#>  [83,]  0.0641600909
#>  [84,]  0.0807674019
#>  [85,]  0.0180828540
#>  [86,] -0.0002938120
#>  [87,]  0.0006463103
#>  [88,]  0.0537719354
#>  [89,]  0.0851661386
#>  [90,]  0.0983258658
#>  [91,]  0.0782045448
#>  [92,]  0.0684126645
#>  [93,]  0.0848222982
#>  [94,] -0.0010883922
#>  [95,] -0.0058721566
#>  [96,]  0.0018729637
#>  [97,]  0.0066104837
#>  [98,]  0.0284932295
#>  [99,] -0.0145268795
#> [100,]  0.0098583588
#> 
#> Slot "weights":
#>   [1] 0.3010606 0.3850167 0.3296584 0.3854821 0.2703865 0.2904376 0.3967625
#>   [8] 0.3896309 0.2975354 0.3834467 0.3506640 0.3660383 0.3861334 0.2781870
#>  [15] 0.3723192 0.3114736 0.3831519 0.3027107 0.3764002 0.3480990 0.2791764
#>  [22] 0.2733036 0.3858744 0.3901590 0.3244229 0.3907559 0.3716687 0.3309954
#>  [29] 0.2709490 0.2717331 0.3953379 0.3832109 0.3686019 0.3745554 0.3862085
#>  [36] 0.3968460 0.3414976 0.3358828 0.3968727 0.3406761 0.3968743 0.3495757
#>  [43] 0.3111759 0.2809656 0.3408453 0.3206091 0.3713728 0.3936252 0.3896931
#>  [50] 0.3130564 0.2804945 0.3237301 0.3951657 0.3696947 0.3261238 0.3700298
#>  [57] 0.3674278 0.3439926 0.3920070 0.3967913 0.2833381 0.2713672 0.3817864
#>  [64] 0.3365415 0.3009117 0.3950104 0.3930144 0.3862627 0.3965723 0.3367357
#>  [71] 0.3893554 0.2841316 0.3969455 0.3816087 0.3823203 0.3963328 0.3303602
#>  [78] 0.3777600 0.3920285 0.3694140 0.3945385 0.3822290 0.3351498 0.3011465
#>  [85] 0.3968180 0.3893043 0.3901301 0.3558121 0.2924465 0.2675060 0.3062967
#>  [92] 0.3263876 0.2931186 0.3885737 0.3836134 0.3911561 0.3943711 0.3918831
#>  [99] 0.3725805 0.3958686
#> 

# generate demand points for a 2d space using hypervolume
dps2 <- make.DemandPoints(points = env.points,
                          kernel.method = "hypervolume",
                          samples.per.point = 50,
                          verbose = FALSE)

# print object
print(dps2)

#> An object of class "DemandPoints"
#> Slot "coords":
#>                 dc1          dc2
#>   [1,]  0.103205458 -0.270759848
#>   [2,] -0.085328496 -0.009744130
#>   [3,]  0.041406147 -0.296846256
#>   [4,]  0.079482103 -0.229604872
#>   [5,]  0.046924831 -0.170428211
#>   [6,] -0.007922185 -0.198939745
#>   [7,] -0.046816510 -0.166434118
#>   [8,]  0.059627949 -0.128834556
#>   [9,] -0.012363344 -0.106882963
#>  [10,] -0.047634302 -0.149027311
#>  [11,] -0.036732667 -0.147474150
#>  [12,]  0.018702909 -0.225276191
#>  [13,] -0.090535026 -0.050881603
#>  [14,]  0.045229929 -0.301520476
#>  [15,] -0.091211087 -0.068792879
#>  [16,]  0.028014395 -0.101737827
#>  [17,]  0.031965494 -0.296232937
#>  [18,] -0.010051068 -0.161969570
#>  [19,] -0.091126689 -0.043993782
#>  [20,]  0.058414496 -0.213916734
#>  [21,]  0.039587248 -0.154114078
#>  [22,]  0.086394162 -0.264011804
#>  [23,]  0.035062273 -0.212467723
#>  [24,]  0.033169515 -0.168449979
#>  [25,]  0.062853461 -0.156789258
#>  [26,]  0.053002849 -0.322839432
#>  [27,]  0.077608412 -0.269519091
#>  [28,]  0.031270681 -0.047853699
#>  [29,] -0.008909874 -0.106109787
#>  [30,] -0.010185851 -0.113369486
#>  [31,] -0.043943918 -0.064707636
#>  [32,]  0.007770390 -0.035588634
#>  [33,] -0.002335254 -0.046421446
#>  [34,] -0.031938743 -0.008315658
#>  [35,] -0.018210636 -0.245074929
#>  [36,] -0.017681526 -0.011978793
#>  [37,] -0.056008957 -0.170724649
#>  [38,]  0.037835993 -0.239396938
#>  [39,]  0.001425408 -0.233868533
#>  [40,]  0.113577726 -0.327378811
#>  [41,]  0.052524278 -0.283522356
#>  [42,] -0.070554523  0.011880192
#>  [43,]  0.076420418 -0.200812962
#>  [44,]  0.004750353 -0.174420334
#>  [45,]  0.030490038 -0.269612692
#>  [46,]  0.045916063 -0.123269871
#>  [47,]  0.061171833 -0.115692147
#>  [48,]  0.032966403 -0.304903415
#>  [49,]  0.023434421 -0.071411571
#>  [50,] -0.071805197 -0.111164958
#>  [51,]  0.091887914 -0.339704401
#>  [52,] -0.003402752 -0.214457728
#>  [53,]  0.069548324 -0.167132126
#>  [54,] -0.065616804 -0.099527492
#>  [55,] -0.051339521 -0.013730011
#>  [56,] -0.023103508 -0.223515411
#>  [57,]  0.021393301 -0.146078421
#>  [58,]  0.037028388 -0.291955925
#>  [59,] -0.062766325 -0.067795300
#>  [60,]  0.040389257 -0.282798928
#>  [61,] -0.047394023 -0.034390859
#>  [62,]  0.037125313 -0.136013060
#>  [63,]  0.041547655 -0.068306700
#>  [64,]  0.089001853 -0.244351304
#>  [65,]  0.034734041 -0.261130602
#>  [66,]  0.073504583 -0.297272345
#>  [67,]  0.051565201 -0.169916592
#>  [68,]  0.042581763 -0.078789625
#>  [69,] -0.025513777 -0.020843036
#>  [70,]  0.052690576 -0.259826501
#>  [71,]  0.008641090 -0.283079441
#>  [72,]  0.051427074 -0.227520244
#>  [73,] -0.069716956 -0.020988682
#>  [74,] -0.028030954 -0.202230001
#>  [75,]  0.039044921 -0.260226545
#>  [76,] -0.038567514 -0.124763341
#>  [77,] -0.009898050 -0.217664188
#>  [78,] -0.034729064 -0.130200631
#>  [79,] -0.060130514 -0.090645702
#>  [80,] -0.031066691 -0.209257352
#>  [81,]  0.046911035 -0.273045693
#>  [82,] -0.048418444 -0.195028312
#>  [83,] -0.048293835 -0.125286629
#>  [84,]  0.070237107 -0.250139757
#>  [85,]  0.081467727 -0.330927654
#>  [86,]  0.052027543 -0.286482892
#>  [87,]  0.008298194 -0.247212157
#>  [88,] -0.078539265 -0.088461963
#>  [89,] -0.058297844 -0.066370019
#>  [90,]  0.045493200 -0.259938937
#>  [91,] -0.057124622 -0.174897773
#>  [92,]  0.032237191 -0.040528186
#>  [93,]  0.001807357 -0.189862705
#>  [94,]  0.004914528 -0.243133339
#>  [95,] -0.052027071 -0.076230645
#>  [96,]  0.090735543 -0.280933800
#>  [97,]  0.097320763 -0.295031475
#>  [98,] -0.088135994 -0.029341170
#>  [99,]  0.031239480 -0.104618267
#> [100,] -0.011279293 -0.221203600
#> 
#> Slot "weights":
#>   [1] 0.0009278472 0.0008758480 0.0010384468 0.0010009476 0.0011975990
#>   [6] 0.0016118394 0.0010909859 0.0008934505 0.0012982197 0.0011500441
#>  [11] 0.0013011059 0.0016101852 0.0008900835 0.0010180540 0.0008509520
#>  [16] 0.0010484553 0.0009598659 0.0015370350 0.0008787217 0.0012158464
#>  [21] 0.0011997489 0.0010740316 0.0015203475 0.0013569598 0.0009491740
#>  [26] 0.0008419924 0.0011688033 0.0008867494 0.0012783817 0.0013092752
#>  [31] 0.0013800470 0.0009685171 0.0010735359 0.0010063038 0.0010179755
#>  [36] 0.0009622440 0.0008926124 0.0014665610 0.0014603041 0.0009196136
#>  [41] 0.0012177179 0.0008401965 0.0009258500 0.0016154343 0.0012549933
#>  [46] 0.0009961421 0.0008436734 0.0008638050 0.0009931797 0.0009981238
#>  [51] 0.0008860927 0.0015858304 0.0009051139 0.0011652173 0.0011174664
#>  [56] 0.0011945911 0.0013239285 0.0010568892 0.0013040512 0.0011767197
#>  [61] 0.0012713409 0.0011278481 0.0008777598 0.0009577597 0.0013489652
#>  [66] 0.0011562201 0.0011356111 0.0008945415 0.0010813568 0.0013255171
#>  [71] 0.0008685165 0.0013445611 0.0011038587 0.0012827224 0.0013581585
#>  [76] 0.0013202984 0.0014825282 0.0013368817 0.0012741926 0.0011670635
#>  [81] 0.0012773555 0.0008990355 0.0012376344 0.0011869904 0.0009314096
#>  [86] 0.0011955999 0.0013669469 0.0010171065 0.0013414787 0.0013506685
#>  [91] 0.0008461502 0.0008554146 0.0016734029 0.0013857907 0.0013701284
#>  [96] 0.0010860910 0.0010544967 0.0009090005 0.0010377478 0.0014298202
#>