Solve RAP object

This function uses Gurobi to find prioritisations using the input parameter and data stored in a RapUnsolved object, and returns a RapSolved object with outputs in it.

Arguments

a
RapUnsolved or RapSolved object.

b
missing to generate solutions using Gurobi. Prioritisations can be specified using logical, numeric, or matrix objects. This may be useful for evaluating the performance of solutions obtained using other software.

verbose
logical should messages be printed during creation of the initial model matrix?.

...
not used.

Value

RapSolved object

Note

This function is used to solve a RapUnsolved object that has all of its inputs generated. The rap function (without lower case 'r') provides a more general interface for generating inputs and outputs.

See also

RapUnsolved, RapSolved.

Examples

# load RapUnsolved object
data(sim_ru)

# solve it using Gurobi
sim_rs <- solve(sim_ru)

#> The rgorubi R package is not installed

#> To access multiple solutions from the solution pool,
#> 	follow these instructions to install the "rgurobi" R package from GitHub:
#> 	http://bit.ly/1RF19XO

#> Warning for adding variables: zero or small (< 1e-13) coefficients, ignored
#> Optimize a model with 30306 rows, 30100 columns and 120000 nonzeros
#> Variable types: 0 continuous, 30100 integer (30100 binary)
#> Coefficient statistics:
#>   Matrix range     [7e-03, 9e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Found heuristic solution: objective 99
#> Presolve removed 329 rows and 15 columns (presolve time = 5s) ...
#> Presolve removed 360 rows and 15 columns
#> Presolve time: 6.77s
#> Presolved: 29946 rows, 30085 columns, 131504 nonzeros
#> Variable types: 0 continuous, 30085 integer (30085 binary)
#> Presolve removed 67 rows and 1 columns
#> Presolved: 29879 rows, 30084 columns, 127194 nonzeros
#> 
#> Presolve removed 29879 rows and 30084 columns
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>        0    1.0000000e+02   0.000000e+00   1.000000e+02      7s
#>     5715    2.0000000e+01   0.000000e+00   0.000000e+00      9s
#>     5715    2.0000000e+01   0.000000e+00   0.000000e+00      9s
#> 
#> Root relaxation: objective 2.000000e+01, 5715 iterations, 2.18 seconds
#> Total elapsed time = 10.09s
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> *    0     0               0      20.0000000   20.00000  0.00%     -   10s
#> 
#> Explored 0 nodes (11023 simplex iterations) in 10.18 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 2: 20 99 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 2.000000000000e+01, best bound 2.000000000000e+01, gap 0.0000%

#> Warning: some species have space.held values less than 0, and are really poor represente

# evaluate manually specified solution using planning unit indices
sim_rs2 <- solve(sim_ru, seq_len(10))

# evaluate manually specifed solution using binary selections
sim_rs3 <- solve(sim_ru, c(rep(TRUE, 10), rep(FALSE, 90)))

#  evaluate multiple manually specified solutions
sim_rs4 <- solve(sim_ru, matrix(sample(c(0, 1), size = 500, replace = TRUE),
                 ncol = 100, nrow = 5))

#> Warning: some species have space.held values less than 0, and are really poor represente